[ADD] 커뮤니티 글 조회/수정/삭제 예외처리 issue: #8

+++ src/boards/boards.controller.ts
@@ -1,4 +1,4 @@
-import { Logger, UploadedFile, UploadedFiles } from '@nestjs/common';
+import { HttpException, HttpStatus, Logger, NotFoundException, Res, UploadedFiles, UseFilters } from '@nestjs/common';
 import { Body, Controller, Delete, Get, Param, Patch, Post, Query, UseInterceptors } from '@nestjs/common';
 import { FilesInterceptor} from '@nestjs/platform-express';
 import { Boards } from './boards.entity';
@@ -7,25 +7,27 @@ import { CreateBoardDto } from './dto/create-board.dto';
 import { UpdateBoardDto } from './dto/update-board.dto';
 import * as AWS from 'aws-sdk';
 import * as multerS3 from 'multer-s3';
-import { create } from 'domain';
 require("dotenv").config();
 
 const s3 = new AWS.S3();
-  
-let S3ImageName=""
 @Controller('boards')
 export class BoardsController {
     constructor(private boardsService: BoardsService){}
 
     @Get() // 커뮤니티 전체 글 조회 / 카테고리별 조회 / 검색어별 조회
     getAllBoards(@Query() query): Promise <Boards[]>{
-        const { category, keyword } = query;
-        return this.boardsService.getAllBoards(category, keyword);
+        const { category, keyword } = query; // @Query()'에서 해당 쿼리문을 받아 query에 저장하고 변수 받아옴
+        const data = this.boardsService.getAllBoards(category, keyword);
+        console.log(data)
+        return data;
     }
 
     @Get('/:boardId') // 커뮤니티 특정 글 조회
-    getBoard(@Param("boardId") boardId: number): Promise <Boards> {
-        return this.boardsService.getBoardById(boardId);
+    async getBoard(@Param("boardId") boardId: number): Promise <Boards> {
+        const board = await this.boardsService.getBoardById(boardId);
+        if(!board)
+            throw new NotFoundException();
+        return board;
     }
 
     @Post() // 커뮤니티 글 작성
@@ -36,7 +38,7 @@ export class BoardsController {
         contentType: multerS3.AUTO_CONTENT_TYPE, 
         acl: 'public-read',
         key: function (request, file, cb) { // files  for문 돌듯이 먼저 실행
-            S3ImageName = `${Date.now().toString()}-${file.originalname}`; // 파일 올리면 해당 파일의 이름을 받아옴 -> S3에 저장되는 이름
+            let S3ImageName = `${Date.now().toString()}-${file.originalname}`; // 파일 올리면 해당 파일의 이름을 받아옴 -> S3에 저장되는 이름
             cb(null, S3ImageName); // 이게 뭘까?            
         }
       })
@@ -45,18 +47,35 @@ export class BoardsController {
         @UploadedFiles() files: Express.Multer.File[], 
         @Body() createBoardDto: CreateBoardDto
     ) {
-        console.log(createBoardDto)
         return this.boardsService.createBoard(files, createBoardDto);
     }
 
     @Patch('/:boardId') // 커뮤니티 글 수정
-    updateBoard(@Param("boardId") boardId: number, @Body() updateBoardDto: UpdateBoardDto) {
+    async updateBoard(@Res() res, @Param("boardId") boardId: number, @Body() updateBoardDto: UpdateBoardDto){
+        const board = await this.boardsService.getBoardById(boardId);
+        if(!board)
+            throw new NotFoundException();
         this.boardsService.updateBoard(boardId, updateBoardDto);
+        return res
+            .status(HttpStatus.OK)
+            .json({
+                success: true,
+                message:'게시글을 수정했습니다'
+            })
     }
 
     @Delete('/:boardId') // 커뮤니티 글 삭제
-    deleteBoard(@Param("boardId") boardId: number){
-        return this.boardsService.deleteBoard(boardId);
+    async deleteBoard(@Res() res, @Param("boardId") boardId: number){
+        const board = await this.boardsService.getBoardById(boardId);
+        if(!board)
+            throw new NotFoundException();
+        this.boardsService.deleteBoard(boardId);
+        return res
+            .status(HttpStatus.OK)
+            .json({
+                success: true,
+                message:'게시글이 삭제되었습니다'
+            })
     }
 
 }
